client: moved navigation components to their own subdirectory
