+++ client/src/App.js
@@ -15,10 +15,10 @@ import ChallengeEvaluations from './challenge/challenge-evaluations';
 import ChallengeSelect from './challenge/challenge-select';
 import CompletedResults from './result/completed-results';
 import CreatePerformance from './performance/create-performance';
-import Login from './shared-components/login';
-import Navbar from './shared-components/navbar';
+import Login from './navigation/login';
+import Navbar from './navigation/navbar';
 import NewPasswordChangeRequest from './password-change-request/new-password-change-request.jsx';
-import NotFound from './shared-components/not-found';
+import NotFound from './navigation/not-found';
 import PendingResults from './result/pending-results';
 import Profile from './profile/profile';
 import PasswordChangeRequest from './password-change-request/password-change-request';
