+++ client/src/navigation/not-found.scss
